+++ src/canisters/frontend/ic-asset/src/asset/content.rs
