+++ src/canisters/frontend/ic-asset/src/asset/content_encoder.rs
