+++ src/canisters/frontend/ic-asset/src/asset/mod.rs
@@ -1,0 +1,3 @@
+mod config;
+mod content;
+mod content_encoder;
