+++ src/canisters/frontend/ic-asset/src/batch_upload/mod.rs
@@ -1,0 +1,4 @@
+mod operations;
+mod plumbing;
+mod retryable;
+mod semaphores;
