+++ src/canisters/frontend/ic-asset/src/batch_upload/operations.rs
