+++ src/canisters/frontend/ic-asset/src/batch_upload/plumbing.rs
