+++ src/canisters/frontend/ic-asset/src/canister_api/list.rs
