+++ src/canisters/frontend/ic-asset/src/canister_api/method_names.rs
