+++ src/canisters/frontend/ic-asset/src/canister_api/mod.rs
