+++ src/canisters/frontend/ic-asset/src/canister_api/protocol.rs
